RAVE 0.524 RC4
- Solucionat bug al mostrar el menu de la llista, i els items seleccionats.
- Solucionat bug amb els controls desplegables i el foco.
- Solucionat bug visual dels controls per editar teclas rapides.
- Agregades les següents opcions :
-- Ocultar tooltips al pasar amb el mouse per sobre.
-- Mostrar el nom del medi actual en la barra del titul.
-- Guardar valors del Brillo, Contrast, Saturació, i Proporció per cada medi.
-- Barrejar llistes aleatories generades per Estil, Grup, Disc, 50Can, Nota

+++ Reproductor de Audio y Video Extendido/DListaEx.cpp
@@ -551,7 +551,7 @@ namespace DWL {
 		}
 
 		// Altura (total de items * altura de la fuente)
-		_TotalAltoVisible = (Fuente.Alto() + (DLISTAEX_PADDING * (int)2)) * _Items.size();
+		_TotalAltoVisible = (Fuente.Alto() + (DLISTAEX_PADDING * 2)) * _Items.size();
 	}
